Update for mac installer (#6)
Update for MacOS installer

I'm just adding support for the Mac installer.

- [x] All existing scripts remains the same.
- [x] Handle git failures gracefully.

Testing : 
Tested by creating an installer on a Mac machine and installing it.  

Note : This will work only on Intel machines. Arm chip (M1, M2) support
has to be added after porta is replaced.

Signing has to be added for seamless installation and launch.